+++ workspace-service/src/graphql/file_download_urls.rs
@@ -1,5 +1,5 @@
 use super::{azure, db};
-use crate::{db::RepoFactory, services::workspace::RepoCreator};
+use crate::{core::RepoCreator, db::RepoFactory};
 use async_graphql::{Context, FieldResult, Object, ID};
 use fnhs_event_models::{Event, EventClient, EventPublisher as _, FileDownloadedData};
 use sqlx::PgPool;
