+++ workspace-service/src/graphql/folders.rs
@@ -1,5 +1,5 @@
 use super::{db, RequestingUser};
-use crate::{db::RepoFactory, services::workspace::RepoCreator};
+use crate::{core::RepoCreator, db::RepoFactory};
 use async_graphql::{Context, FieldResult, InputObject, Object, SimpleObject, ID};
 use fnhs_event_models::{
     Event, EventClient, EventPublisher, FolderCreatedData, FolderDeletedData, FolderUpdatedData,
