+++ workspace-service/src/services/mod.rs
@@ -1,3 +1,1 @@
-pub mod team;
-pub mod user;
 pub mod workspace;
